Started working on fixes for macOS
QT has a bug where a full screen window cannot be hidden properly so I started working on a new version that uses a CocoaWindow that can be hidden properly. I didn't worry about preserving compatibility with Windows because this was meant just for my church. I stopped work on it when ProPresenter 7 was announced which has support for Dual Language Bible Display so my effort here is unnecessary.